refactor(lance): keep parser APIs out of root
