+++ examples/lance/browser/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "lance-examples",
+  "name": "lance-browser-example",
   "version": "0.0.0",
   "private": true,
   "type": "module",
