+++ modules/lance/src/index.ts
@@ -8,37 +8,7 @@ export type {LanceLoaderOptions} from './lance-loader-types';
 export {LanceSourceLoader, LanceSource} from './lance-source-loader';
 export type {LanceSourceLoaderOptions, LanceSourceMetadata} from './lance-source-loader';
 export {
-  parseLanceManifest,
-  type LanceManifest,
-  type LanceManifestDataFile,
-  type LanceManifestField,
-  type LanceManifestFragment
-} from './lance-manifest';
-export {
-  parseLanceFileMetadata,
-  parseLanceColumnMetadata,
-  type LanceFileMetadata,
-  type LanceFileColumnMetadata,
-  type LanceFilePageMetadata,
-  type LanceFileRange
-} from './lance-file';
-export {
-  decodeLanceFlatColumn,
-  decodeLanceFlatPage,
-  LanceFlatPageUnsupportedError,
-  type LanceFlatPrimitiveArray,
-  type LanceFlatPrimitiveType
-} from './lance-decoder';
-export {
-  parseLanceFileToArrow,
-  readLanceRemoteCoordinatesToArrow,
-  readLanceRemoteFileToArrow,
-  type LanceRemoteCoordinateRead,
-  type LanceRemoteColumnRead,
-  type LanceArrowReadOptions
-} from './lance-arrow';
-export {
   LANCE_SOURCE_CAPABILITIES,
   type LanceSourceCapabilities
 } from './lance-source-capabilities';
-export {LanceDecoderUnavailableError} from './lance-loader';
+export {LanceDecoderUnavailableError} from './lance-errors';
