+++ modules/lance/src/unbundled.ts
@@ -6,34 +6,4 @@ export {LanceLoader} from './lance-loader-types';
 export type {LanceLoaderOptions} from './lance-loader-types';
 export {LanceSourceLoader, LanceSource} from './lance-source-loader';
 export type {LanceSourceLoaderOptions, LanceSourceMetadata} from './lance-source-loader';
-export {LanceDecoderUnavailableError} from './lance-loader';
-export {
-  parseLanceManifest,
-  type LanceManifest,
-  type LanceManifestDataFile,
-  type LanceManifestField,
-  type LanceManifestFragment
-} from './lance-manifest';
-export {
-  parseLanceFileMetadata,
-  parseLanceColumnMetadata,
-  type LanceFileMetadata,
-  type LanceFileColumnMetadata,
-  type LanceFilePageMetadata,
-  type LanceFileRange
-} from './lance-file';
-export {
-  decodeLanceFlatColumn,
-  decodeLanceFlatPage,
-  LanceFlatPageUnsupportedError,
-  type LanceFlatPrimitiveArray,
-  type LanceFlatPrimitiveType
-} from './lance-decoder';
-export {
-  parseLanceFileToArrow,
-  readLanceRemoteCoordinatesToArrow,
-  readLanceRemoteFileToArrow,
-  type LanceRemoteCoordinateRead,
-  type LanceRemoteColumnRead,
-  type LanceArrowReadOptions
-} from './lance-arrow';
+export {LanceDecoderUnavailableError} from './lance-errors';
